+++ packages/gatsby/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.32.4](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.32.3...gatsby@2.32.4) (2021-02-18)
+
+### Bug Fixes
+
+- drop terminal-link ([#29472](https://github.com/gatsbyjs/gatsby/issues/29472)) ([#29477](https://github.com/gatsbyjs/gatsby/issues/29477)) ([6374419](https://github.com/gatsbyjs/gatsby/commit/637441942806a439cc7efc4c7ba54be24c59d276))
+
 ## [2.32.3](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.32.2...gatsby@2.32.3) (2021-02-05)
 
 **Note:** Version bump only for package gatsby